Move SilProposal into data layer
